requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
beautifulsoup4>=4.12.2
openai>=1.3.0
tqdm>=4.66.1
pytest>=7.0.0